refactor(core): desktop project struct (#8334)

+++ packages/frontend/core/src/components/page-list/operation-cell.tsx
@@ -4,11 +4,11 @@ import {
   MenuItem,
   toast,
   useConfirmModal,
+  usePromptModal,
 } from '@affine/component';
 import { useBlockSuiteMetaHelper } from '@affine/core/components/hooks/affine/use-block-suite-meta-helper';
-import { useTrashModalHelper } from '@affine/core/components/hooks/affine/use-trash-modal-helper';
 import { useCatchEventCallback } from '@affine/core/components/hooks/use-catch-event-hook';
-import { DocInfoService } from '@affine/core/modules/doc-info';
+import { WorkspaceDialogService } from '@affine/core/modules/dialogs';
 import {
   CompatibleFavoriteItemsAdapter,
   FavoriteService,
@@ -33,6 +33,7 @@ import {
   SplitViewIcon,
 } from '@blocksuite/icons/rc';
 import {
+  DocsService,
   FeatureFlagService,
   useLiveData,
   useService,
@@ -51,7 +52,6 @@ import { DisablePublicSharing, MoveToTrash } from './operation-menu-items';
 import { CreateOrEditTag } from './tags/create-tag';
 import type { TagMeta } from './types';
 import { ColWrapper } from './utils';
-import { useEditCollection, useEditCollectionName } from './view';
 
 const tooltipSideTop = { side: 'top' as const };
 const tooltipSideTopAlignEnd = { side: 'top' as const, align: 'end' as const };
@@ -83,17 +83,19 @@ export const PageOperationCell = ({
     featureFlagService.flags.enable_multi_view.$
   );
   const currentWorkspace = workspaceService.workspace;
-  const { setTrashModal } = useTrashModalHelper();
   const favourite = useLiveData(favAdapter.isFavorite$(page.id, 'doc'));
   const workbench = workbenchService.workbench;
   const { duplicate } = useBlockSuiteMetaHelper();
+  const docRecord = useLiveData(useService(DocsService).list.doc$(page.id));
   const blocksuiteDoc = currentWorkspace.docCollection.getDoc(page.id);
 
-  const docInfoModal = useService(DocInfoService).modal;
+  const workspaceDialogService = useService(WorkspaceDialogService);
   const onOpenInfoModal = useCallback(() => {
-    track.$.docInfoPanel.$.open();
-    docInfoModal.open(blocksuiteDoc?.id);
-  }, [blocksuiteDoc?.id, docInfoModal]);
+    if (blocksuiteDoc?.id) {
+      track.$.docInfoPanel.$.open();
+      workspaceDialogService.open('doc-info', { docId: blocksuiteDoc.id });
+    }
+  }, [blocksuiteDoc?.id, workspaceDialogService]);
 
   const onDisablePublicSharing = useCallback(() => {
     // TODO(@EYHN): implement disable public sharing
@@ -102,15 +104,26 @@ export const PageOperationCell = ({
     });
   }, []);
 
+  const { openConfirmModal } = useConfirmModal();
+
   const onRemoveToTrash = useCallback(() => {
+    if (!docRecord) {
+      return;
+    }
     track.allDocs.list.docMenu.deleteDoc();
 
-    setTrashModal({
-      open: true,
-      pageIds: [page.id],
-      pageTitles: [page.title],
+    openConfirmModal({
+      title: t['com.affine.moveToTrash.confirmModal.title'](),
+      description: t['com.affine.moveToTrash.confirmModal.description']({
+        title: docRecord.title$.value || t['Untitled'](),
+      }),
+      cancelText: t['com.affine.confirmModal.button.cancel'](),
+      confirmText: t.Delete(),
+      onConfirm: () => {
+        docRecord.moveToTrash();
+      },
     });
-  }, [page.id, page.title, setTrashModal]);
+  }, [docRecord, openConfirmModal, t]);
 
   const onOpenInSplitView = useCallback(() => {
     track.allDocs.list.docMenu.openInSplitView();
@@ -297,11 +310,15 @@ export const CollectionOperationCell = ({
   info,
 }: CollectionOperationCellProps) => {
   const t = useI18n();
-  const { compatibleFavoriteItemsAdapter: favAdapter, workspaceService } =
-    useServices({
-      CompatibleFavoriteItemsAdapter,
-      WorkspaceService,
-    });
+  const {
+    compatibleFavoriteItemsAdapter: favAdapter,
+    workspaceService,
+    workspaceDialogService,
+  } = useServices({
+    CompatibleFavoriteItemsAdapter,
+    WorkspaceService,
+    WorkspaceDialogService,
+  });
   const docCollection = workspaceService.workspace.docCollection;
   const { createPage } = usePageHelper(docCollection);
   const { openConfirmModal } = useConfirmModal();
@@ -309,11 +326,7 @@ export const CollectionOperationCell = ({
     favAdapter.isFavorite$(collection.id, 'collection')
   );
 
-  const { open: openEditCollectionModal } = useEditCollection();
-
-  const { open: openEditCollectionNameModal } = useEditCollectionName({
-    title: t['com.affine.editCollection.renameCollection'](),
-  });
+  const { openPromptModal } = usePromptModal();
 
   const handlePropagation = useCallback((event: MouseEvent) => {
     event.preventDefault();
@@ -323,39 +336,36 @@ export const CollectionOperationCell = ({
   const handleEditName = useCallback(
     (event: MouseEvent) => {
       handlePropagation(event);
-      // use openRenameModal if it is in the sidebar collection list
-      openEditCollectionNameModal(collection.name)
-        .then(name => {
-          return service.updateCollection(collection.id, collection => ({
+      openPromptModal({
+        title: t['com.affine.editCollection.renameCollection'](),
+        label: t['com.affine.editCollectionName.name'](),
+        inputOptions: {
+          placeholder: t['com.affine.editCollectionName.name.placeholder'](),
+        },
+        confirmText: t['com.affine.editCollection.save'](),
+        cancelText: t['com.affine.editCollection.button.cancel'](),
+        confirmButtonOptions: {
+          variant: 'primary',
+        },
+        onConfirm(name) {
+          service.updateCollection(collection.id, () => ({
             ...collection,
             name,
           }));
-        })
-        .catch(err => {
-          console.error(err);
-        });
+        },
+      });
     },
-    [
-      collection.id,
-      collection.name,
-      handlePropagation,
-      openEditCollectionNameModal,
-      service,
-    ]
+    [collection, handlePropagation, openPromptModal, service, t]
   );
 
   const handleEdit = useCallback(
     (event: MouseEvent) => {
       handlePropagation(event);
-      openEditCollectionModal(collection)
-        .then(collection => {
-          return service.updateCollection(collection.id, () => collection);
-        })
-        .catch(err => {
-          console.error(err);
-        });
+      workspaceDialogService.open('collection-editor', {
+        collectionId: collection.id,
+      });
     },
-    [handlePropagation, openEditCollectionModal, collection, service]
+    [handlePropagation, workspaceDialogService, collection.id]
   );
 
   const handleDelete = useCallback(() => {
